+++ NEWS.md
@@ -1,3 +1,8 @@
+# Current
+
+* Rename hsEventsOnChange() to eventsOnChange() and use new function
+  kwb.utils::findChanges() within the function
+
 # kwb.event 0.3.0
 
 * Added a `NEWS.md` file to track changes to the package.